Fix to clutter objects spawing as wands when links are < 256 Fix to give_ptr_npc to update object pointers after the object list is updated Fix the UW1 chair model material Set x_skills for UW1 to give only 6 point to a skill in mode 10001 Capped skill gains to 30. Fixed using other objects on the pole. Stopped unlinked spell traps form spawning and occupying entire object lists. Fixed Impact object no of frames Stopped picking up of Orbs Stopped owner from being cleared when value >=31 Fixed @GI string replacement.

+++ UnityScripts/scripts/Loaders/ArtLoader.cs
@@ -12,7 +12,7 @@ public class ArtLoader : Loader {
 	/// </summary>
 	protected bool LoadMod; 
 
-	public	const int BitMapHeaderSize=28;
+	public const byte BitMapHeaderSize=28;
 
 	/// <summary>
 	/// The complete image file 
@@ -22,7 +22,7 @@ public class ArtLoader : Loader {
 		/// <summary>
 		/// The palette no to use with this file.
 		/// </summary>
-	public int PaletteNo=0;
+	public short PaletteNo=0;
 
 	/// <summary>
 	/// Loads the image file into the buffer
